created "Moves" component with the history management logic
